requirements.txt
pymupdf
pdfplumber
pytesseract
pillow
opencv-python
langchain
sentence-transformers
pydantic
python-dotenv
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml 
pypdf
bs4
arxiv
wikipedia
langchain_huggingface
nltk
scikit-learn
opensearch-py
faiss-cpu
langchain-google-genai
python-docx
langgraph
streamlit
ipython
langchain_tavily